backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
motor==3.1.1
pymongo==4.5.0
websockets==12.0